requirements.txt
fastmcp
openai
fastapi
uvicorn[standard]
ollama
pyyaml
